requirements.txt
# PDF to PPT Converter - Python Dependencies

# PDF parsing
pymupdf>=1.23.0

# PPT generation
python-pptx>=1.0.0

# Web framework
fastapi>=0.115.0
uvicorn>=0.32.0
jinja2>=3.1.0
python-multipart>=0.0.9

# CLI
click>=8.1.0

# OCR support
pytesseract>=0.3.10
Pillow>=10.0.0
